revert to 1:1 file names
