+++ AltStoreCore/Roxas/NSFileManager+URLs.swift
@@ -1,8 +1,8 @@
 //
-//  FileManager+URLs.swift
+//  NSFileManager+URLs.swift
 //  AltStoreCore
 //
-//  Created by Magesh K on 06/17/26.
+//  Created by Magesh K on 6/17/26.
 //  Copyright © 2026 SideStore. All rights reserved.
 //
