+++ AltStoreCore/Roxas/UICollectionViewCell+CellContent.swift
@@ -1,0 +1,12 @@
+//
+//  UICollectionViewCell+CellContent.swift
+//  AltStoreCore
+//
+//  Created by Magesh K on 6/17/26.
+//  Copyright © 2026 SideStore. All rights reserved.
+//
+
+import UIKit
+extension UICollectionViewCell: RSTCellContentCell {}
+
+
