+++ AltStoreCore/Roxas/UIView+AnimatedHide.swift
@@ -1,8 +1,8 @@
 //
-//  UIView+Extensions.swift
+//  UIView+AnimatedHide.swift
 //  AltStoreCore
 //
-//  Created by Magesh K on 06/17/26.
+//  Created by Magesh K on 6/17/26.
 //  Copyright © 2026 SideStore. All rights reserved.
 //
